requirements.txt
# EdgeVLM Dependencies

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# CLI
click>=8.1.0
rich>=13.7.0

# ML/AI
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
numpy>=1.24.0  # For LiDAR point cloud processing

# Database
asyncpg>=0.29.0  # PostgreSQL async driver

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0  # Configuration management

# Configuration parsing
pyyaml>=6.0.0  # For YAML config files

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

